pyproject.toml
[project]
name = "gtapilot"
version = "0.1.0"
description = ""
authors = [{ name = "Michael Wang", email = "41971562+devmwang@users.noreply.github.com" }]
requires-python = "~=3.13"
readme = "README.md"
dependencies = [
    "bettercam==1.0.0",
    "comtypes==1.4.12",
    "cuda-python==12.9.2",
    "numpy==2.3.2",
    "opencv-python==4.11.0.86",
    "pyzmq==27.0.2",
    "setproctitle==1.3.6",
    "tensorrt==10.13.0.35",
    "tensorrt-stubs==10.13.0.35.1",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "transformers>=4.56.1",
]

[project.urls]
Repository = "https://github.com/devmwang/GTAP"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[[tool.uv.index]]
name = "pytorch-cu129"
url = "https://download.pytorch.org/whl/cu129"
explicit = true

[tool.uv.sources]
torch = [
  { index = "pytorch-cu129", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cu129", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
